requirements.txt

transformers>=4.41
torch
Pillow
pydantic>=2
pyyaml
langchain
langchain-community
# Optional for big model fallback with Ollama:
# langchain-ollama
